requirements.txt
flask>=3.0.0
google-genai>=0.8.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-flask>=1.3.0
flake8>=6.0.0
black>=23.0.0
